Set labels to all lowercase for standardization
All labels should be lowercase, because it
separates two concerns:

If the label is set to both uppercase and
lowercase, it essentially stores two pieces
of data: The type of the piece, and the color
of the piece. We are already storing the color
of the piece through the faction member variable,
which makes storing color data as upper-lowercase
values duplicate work.