Rename ChessTests.py to FideTests.py
This file in particular checks to see whether
the program satisfies the FIDE rules of chess.

The issue with having this test suite be the
main test suite, is that the rules of chess
are complex, and should only be applied to
a working engine.

Right now, the engine is not working as intended,
and requires a lot of work to even start passing
a single FIDE Test. As a result, it is more
instructive to have a dedicated test suite for
FIDE tests, and a test suite for developing the
engine.