Implement naive Chessboard.toFEN_placement
This commit makes EngineTests.test_1 pass,
however, it is not a complete implementation.

It suffers from the issue that, it will
successfully count the amount of spaces
for boards where pieces never have spaces
to the left of them.

To test this, check if the FEN of this engine
matches the actual FEN after move 1.e4