Rename move_verifier.py to move_generator.py
This is done to split generation logic, and
filtration logic, into two different classes.
One class, the generator, creates moves based on
the Algebreic Notation that's passed in (such as "Ra3")
The other class, the filter, filters a list of
moves based on certain criteria, like whether the
move leaves the board in a state of check.