Completely revert to a time before mypy.
Mypy, static type checking, all this sort of stuff?
1 out of 5 stars. There is definately a time and place
for mypy, static type checking, and all this sort of stuff.
Perhaps if I had used type annotations from the start,
or if the codebase was in a better architectural state,
it would work a lot better.

However, right now, implementing mypy has increased
the amount of code smells, reduced the readability
and evolvablity of the code, and for the love of
god it does not even type check on runtime- leading
to very confusing error messages.

Perhaps this is a rage quit moment- and a more
eloquent commit message can be written in a blog post
or something along those lines later, but right now,
it just does not seem worth it, for this project,
to use static type checking.

It's time to take a nap istg. Maybe I'll return
to mypy later, but right now, it's not the tool
for the job, at all.