Revert 'Completely revert to a time before mypy'
Several important and helpful changes were made
along the way towards an implementation of mypy.
Most of them were combined in the commits that
implemented mypy, which means that it is
better to manually remove the implementation
of mypy, than it is to git revert to a time
before it.

This is a lesson on writing small, concise commits
that does one thing at a time.

At the same time, mypy isn't all that bad-
many of the "mandatory changes" done to
"feed the mypy gods" such as None checks
and other issues, were code smells that mypy
revealed within the program, not just
smells that mypy enforced upon the
codebase.

All in all, after taking a proper rest,
without a headache, I've realized that
we should probably respect the mypy gods.